CHORE: Test fftw3
FFTW3 gives weird DFT outputs. Expected the output to be between 0 and
the size of the fft_size, but the output also goes below 0. The output
gives a complex number back, but neither of them is the phase or
magnitude it seems. I don't know what is really happnening.

Maybe better to build a simple DFT class yourself after looking just a
little further in the docs, but it seems that this fftw library is too
specialized, with to many bells and whistles to be useful for a simple
prototype.